feat: update generating algolia search index for docs
